scp-web/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
opencv-python>=4.9.0
scipy>=1.12.0
pycryptodome>=3.20.0
pymongo>=4.6.0
scikit-learn>=1.4.0
cryptography>=42.0.0
openpyxl>=3.1.0
reportlab>=4.1.0
celery>=5.3.0
redis>=5.0.0